Full Support for wIndows softphone and some refactoring
Refactored some code to move the settings to a function.
Added full support for the windows softphone including automatic updates by using a .appinstaller instead of the .appxbundle directly. The ms-installer was really picky with the file paths and file names and didn't work fully until the file name on the server matched the filename, so I had to move that hook into index.php