Support for Windows Softphone with AppInstaller
You can also update the UpdateInterval because it "proxies" the appinstaller file by importing the one at the URL and serving it to the client when requested. It would be trivial to manipulate/add other attributes if required, but we need the original .appinstaller to be imported so that it matches the version number with the .appxbundle file that is reverenced in the <MainBundle> section. I previously  dynamically generated the XML from scratch.